+++ src/util.js
@@ -11,7 +11,6 @@ export const freeze = Object.freeze ? a => Object.freeze(a) : identity;
 export const emptyArray = freeze([]);
 export const emptyObject = freeze({});
 
-
 export function shallowCopyObject(obj) {
   const copy = {};
   for (let key in obj) {
@@ -22,11 +21,18 @@ export function shallowCopyObject(obj) {
   return copy;
 }
 
-
 export function shallowCopyArray(arr) {
   const copy = Array(arr.length);
   for (let i = 0, len = arr.length; i < len; i++) {
     copy[i] = arr[i];
   }
   return copy;
+}
+
+export function createMessagesWithJsonPath(messages, jsonPath) {
+  return messages.map(message => {
+    const messageCopy = shallowCopyObject(message);
+    messageCopy.path = jsonPath;
+    return messageCopy;
+  });
 }